Don't Detach During Calibrate Chain Lengths
Detach is handled by the axis class and really shouldn't be used
outside of that.

The problem was that calibrate chains was spooling out more chain
than it was supposed to, while reporting that it had done less
that it was supposed to.

This was caused by calling detach inside the calibrate chain length
function.  This caused the axis to stop moving before the endMove
function had completed.  As a result, the location was listed as
being short of the desired termination point, but since the axis
was agressively stopped, its own momemtum would carry it past
where it was supposed to go.

Solution was to remove detach call.

Also changed print statement about the axis location, if we wanted
to read the axis, we would have to block until the axis settled
first.  There is no need for this, we can report what the commanded
length was.